Reorganize and fix tests
