+++ services/app/src/actions/Card.tsx
@@ -4,7 +4,6 @@ import {getNavigator} from '../Globals';
 import {AppStateWithHistory, CardName, CardPhase} from '../reducers/StateTypes';
 import {getStore} from '../Store';
 import {NavigateAction, remoteify} from './ActionTypes';
-import {logQuestPlay} from './Web';
 
 interface ToCardArgs {
   keySuffix?: string;
@@ -39,12 +38,6 @@ export const toCard = remoteify(function toCard(a: ToCardArgs, dispatch: Redux.D
   }
   if (a.keySuffix) {
     keylist.push(a.keySuffix);
-  }
-
-  // Log the end of the quest if we reach it.
-  if (a.name === 'QUEST_END') {
-    console.log('Dispatching logQuestPlay');
-    dispatch(logQuestPlay({phase: 'end'}));
   }
 
   dispatch({type: 'NAVIGATE', to: {...a, ts: Date.now(), key: keylist.join('|'), questId}, dontUpdateUrl: state.settings && state.settings.simulator} as NavigateAction);
