+++ services/app/src/actions/QuestHistory.test.tsx
@@ -1,0 +1,3 @@
+describe('QuestHistory', () => {
+  // Entirely glue code
+});
