+++ services/cards/webpack.config.js
@@ -31,7 +31,7 @@ module.exports = {
   },
   module: {
     rules: [
-      { test: /\.tsx$/, enforce: 'pre', loader: 'tslint-loader' },
+      { test: /\.tsx$/, enforce: 'pre', loader: 'tslint-loader', options: {fix: true} },
       { test: /\.(ttf|eot|svg|png|gif|jpg|woff(2)?)(\?[a-z0-9=&.]+)?$/, loader : 'file-loader' },
       { test: /\.scss$/, loader: 'style-loader!css-loader!sass-loader' },
       { test: /\.json$/, loader: 'json-loader' },
